break up router.js, reorganize into /router
