+++ src/router/create-routes.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Route } from "react-router";
-import Home from "./components/pages/home";
+import Home from "../components/pages/home";
 
 export default function createRoutes(){
   return (
